news_pipeline/requirement.txt
jsonrpclib-pelix==0.3.1
pika==0.11.2
requests==2.18.4
redis==2.10.6
sklearn==0.0
python-dateutil==2.6.0
newspaper3k==0.2.6